brazilian translation plus request
Hi, i did this translation for am2r, it's fully done and i reviewed through several dictionaries. I wanted to report that the GUI needs to be redone because most of times the portuguese words are bigger than the words in english language, the GUI cant fully display the words/phrases in several places.

Also, the fonts are hard to see, it would be good if the fonts didnt have jagged borders.